Implement Stats in terms of AnonStats
Stats now simply provide names for the index based counters contained
in an embedded AnonStats. This removes the use of mutex in favour of
the sync/atomic based approach of AnonStats which is faster and avoids
unnecessary synchronization between goroutines.

CPU profiling for the filter and writer used to feature stats update
and synchronization operations but these no longer show up.

The benchmark in the stats package is now ~78% faster.

+++ filter/filter.go
@@ -208,7 +208,7 @@ func (f *Filter) updateNATSDropped(st *stats.Stats) {
 		log.Printf("NATS: failed to read subscription drops: %v", err)
 		return
 	}
-	st.Max(statNATSDropped, dropped)
+	st.Max(statNATSDropped, uint64(dropped))
 }
 
 func initStats() *stats.Stats {
